Fix pipeline crash: upscale/denoise deferral to FFmpeg
- Remove Upscale exclusion from backend-eligible stages so it routes
  through MiGraphX like other stages
- Fix encodeWithAiProcessing: when backend is null or defers, the
  fallback FFmpeg filter path now always creates the temp file
  (was leaving currentVideoPath pointing at a nonexistent temp_0.mkv)
- Remove dead encodeWithAiUpscale() code (was inside #if 0)
- Add Vulkan runtime, GLSL shader backend, VapourSynth backend,
  filter catalog, and filter browser stubs
- Extend GUI with model manager and toggle switch improvements

+++ src/gui/toggle_switch.cpp
@@ -72,7 +72,7 @@ void ToggleSwitch::paintEvent(QPaintEvent* /*e*/) {
     track.setRedF(  trackFrom.redF()   + (trackTo.redF()   - trackFrom.redF())   * thumbPos_);
     track.setGreenF(trackFrom.greenF() + (trackTo.greenF() - trackFrom.greenF()) * thumbPos_);
     track.setBlueF( trackFrom.blueF()  + (trackTo.blueF()  - trackFrom.blueF())  * thumbPos_);
-    track.setAlphaF(hovered_ ? 0.95 : 0.85);
+    track.setAlphaF(static_cast<float>(hovered_ ? 0.95 : 0.85));
 
     const QRectF trackRect(0, (height() - kTrackH) / 2.0, kTrackW, kTrackH);
     QPainterPath trackPath;
